Moved TopicMessageDeliveryStrategy to config utils.
